backend/requirements.txt
fastapi==0.75.0
uvicorn==0.17.6
together
python-dotenv
openai